% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ColSumtwomatrix_C}
\alias{ColSumtwomatrix_C}
\title{ColSumtwomatrix_C}
\usage{
ColSumtwomatrix_C(AA, BB)
}
\arguments{
\item{AA}{Matrix}

\item{BB}{Matrix}
}
\description{
ColSumtwomatrix_C
}
\keyword{internal}
