% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{doubleloop_LRT}
\alias{doubleloop_LRT}
\title{doubleloop_LRT}
\usage{
doubleloop_LRT(K1R, K2R, P0R, AR, U1R, wR, LamdasR, nminuspx, all_rho,
  LR0_allRhoR)
}
\arguments{
\item{K1R}{K1 matrix}

\item{K2R}{K2 matrix}

\item{P0R}{P0 matrix}

\item{AR}{A matrix}

\item{U1R}{U1 vector}

\item{wR}{w matrix}

\item{LamdasR}{Lamdas vector}

\item{nminuspx}{n-px}

\item{all_rho}{rho vector}

\item{LR0_allRhoR}{LR0_allRhomatrix}
}
\description{
doubleloop_LRT
}
\keyword{internal}
