% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getp_au1.R
\name{getp_au1}
\alias{getp_au1}
\title{Title}
\usage{
getp_au1(null, LR)
}
\arguments{
\item{null}{a vector of likelihood ratio under the null hypothesis. Generated from simulation}

\item{LR}{the likelihood of the data}
}
\description{
Title
}
\keyword{internal}
