% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getp_aud_estimate_pi_first.R
\name{getp_aud_estimate_pi_first}
\alias{getp_aud_estimate_pi_first}
\title{Title}
\usage{
getp_aud_estimate_pi_first(null, LR)
}
\arguments{
\item{null}{a vector of likelihood ratio under the null hypothesis. Generated from simulation}

\item{LR}{the likelihood of the data}
}
\description{
Title
}
\keyword{internal}
