% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getp_Qreg_logP.R
\name{getp_Qreg_logP}
\alias{getp_Qreg_logP}
\title{Title}
\usage{
getp_Qreg_logP(null, LR, qmax = 1)
}
\arguments{
\item{null}{a vector of likelihood ratio under the null hypothesis. Generated from simulation}

\item{LR}{the likelihood of the data}

\item{qmax}{quantile of likelihood ratio vector. Most time put as 1, all the likelihood ratio vector under the null hypothesis will be used}
}
\description{
Title
}
\keyword{internal}
