% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LR0_fixRho_C}
\alias{LR0_fixRho_C}
\title{LR0_fixRho_C}
\usage{
LR0_fixRho_C(LamdasR, muR, w1R, w2R, nminuspx)
}
\arguments{
\item{LamdasR}{Lamda Number}

\item{muR}{mu vector}

\item{w1R}{w1 vector}

\item{w2R}{w2 vector}

\item{nminuspx}{n-px}
}
\description{
LR0_fixRho_C
}
\keyword{internal}
