% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LR0_fixRho_LRT_C}
\alias{LR0_fixRho_LRT_C}
\title{LR0_fixRho_LRT_C}
\usage{
LR0_fixRho_LRT_C(LamdasR, muR, w1R, w2R, nminuspx, xiR)
}
\arguments{
\item{LamdasR}{Lamda Number}

\item{muR}{mu vector}

\item{w1R}{w1 vector}

\item{w2R}{w2 vector}

\item{nminuspx}{n-px}

\item{xiR}{Vector}
}
\description{
LR0_fixRho_LRT_C
}
\keyword{internal}
