% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omniLRT_fast.R, R/omniRLRT_fast.R
\name{omniRLRT_fast}
\alias{omniRLRT_fast}
\alias{omniRLRT_fast}
\title{Title}
\usage{
omniRLRT_fast(y, X, K1, K2, N = 10000, length.rho = 200,
  length.lambda = 21)

omniRLRT_fast(y, X, K1, K2, N = 10000, length.rho = 200,
  length.lambda = 21)
}
\arguments{
\item{length.lambda}{}

\item{length.lambda}{}
}
\description{
Title

Title
}
