% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onmiLRT_fast.R
\name{omniLRT_fast}
\alias{omniLRT_fast}
\title{Title}
\usage{
omniLRT_fast(y, X, K1, K2, N = 10000, N.aud = 1000, length.lambda = 200,
  length.rho = 21)
}
\arguments{
\item{length.rho}{}
}
\description{
Title
}
